move transformer scripts to transformers modules
