+++ src/diffusers/models/unets/unet_3d_blocks.py
@@ -20,7 +20,6 @@ from torch import nn
 from ...utils import is_torch_version
 from ...utils.torch_utils import apply_freeu
 from ..attention import Attention
-from ..dual_transformer_2d import DualTransformer2DModel
 from ..resnet import (
     Downsample2D,
     ResnetBlock2D,
@@ -28,8 +27,9 @@ from ..resnet import (
     TemporalConvLayer,
     Upsample2D,
 )
-from ..transformer_2d import Transformer2DModel
-from ..transformer_temporal import (
+from ..transformers.dual_transformer_2d import DualTransformer2DModel
+from ..transformers.transformer_2d import Transformer2DModel
+from ..transformers.transformer_temporal import (
     TransformerSpatioTemporalModel,
     TransformerTemporalModel,
 )
